backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.35
asyncpg==0.29.0
redis==5.1.0
httpx==0.27.2
faiss-cpu==1.8.0.post1
python-multipart==0.0.9
